+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - MBean name of ClasspathAgent changed to "clazzfish:type=agent,agent=ClasspathAgent"
+- MBeanHelper class replaced by MBeanFinder class
 - update to Java 17
 
 ### Added